--- a/.changelog/2166.internal.md
+++ b/.changelog/2166.internal.md
@@ -1,0 +1,1 @@
+Speedup playwright CI