--- a/.github/workflows/ci-test.yml
+++ b/.github/workflows/ci-test.yml
@@ -56,10 +56,36 @@ jobs:
           path: coverage
           retention-days: 5
 
+  playwright-prepare:
+    # NOTE: This name appears in GitHub's Checks API.
+    name: playwright-prepare
+    needs: [yarn_cache]
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+      - name: Set up Node.js 18
+        uses: actions/setup-node@v4
+        with:
+          node-version: '18.x'
+          cache: yarn
+      - name: Install dependencies
+        run: yarn install --frozen-lockfile
+      - run: REACT_APP_E2E_TEST=1 yarn build:ext
+      - run: REACT_APP_E2E_TEST=1 yarn build
+      - name: Upload builds for playwright
+        uses: actions/upload-artifact@v4
+        with:
+          name: builds-for-playwright
+          path: |
+            ./build
+            ./build-ext
+          retention-days: 5
+
   playwright:
     # NOTE: This name appears in GitHub's Checks API.
     name: playwright
-    needs: [yarn_cache]
+    needs: [playwright-prepare]
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -74,15 +100,15 @@ jobs:
         with:
           node-version: '18.x'
           cache: yarn
-      # TODO: could move outside sharding: yarn + build:ext + build + upload
-      # and then 8x shard download builds + serve + test
-      - name: Install dependencies
-        run: yarn install --frozen-lockfile
-      - run: REACT_APP_E2E_TEST=1 yarn build:ext
-      - run: REACT_APP_E2E_TEST=1 yarn start:prod &
+      - uses: actions/download-artifact@v4
+      - run: mv ./builds-for-playwright/build ./build
+      - run: mv ./builds-for-playwright/build-ext ./build-ext
       - name: Install playwright's npm dependencies
         working-directory: ./playwright/
         run: yarn install --frozen-lockfile
+      # outside sharding: yarn + build:ext + build + upload
+      # and then 8x shard: download builds + serve + test
+      - run: REACT_APP_E2E_TEST=1 node ./playwright/serve-prod.js &
       - name: Install playwright's system dependencies
         working-directory: ./playwright/
         run: npx playwright install --with-deps