--- a/src/config.ts
+++ b/src/config.ts
@@ -1,4 +1,3 @@
-import { Capacitor } from '@capacitor/core'
 import { NetworkType } from 'app/state/network/types'
 
 export const TRANSACTIONS_LIMIT = 20
@@ -197,13 +196,6 @@ export const paraTimesConfig: ParaTimesConfig = {
   [ParaTime.Emerald]: emeraldConfig,
   [ParaTime.Sapphire]: sapphireConfig,
 }
-
-// https://github.com/mozilla/webextension-polyfill/blob/6e3e26c/src/browser-polyfill.js#L9
-export const runtimeIs = Capacitor.isNativePlatform()
-  ? 'mobile-app'
-  : (globalThis as any).chrome?.runtime?.id
-  ? 'extension'
-  : 'webapp'
 
 const splitUrls = (input: string | undefined): string[] =>
   input