Before: [5m 47s](https://github.com/oasisprotocol/wallet/actions/runs/14242574836/) (billed time 7m 1s)
Shard: [2m 53s](https://github.com/oasisprotocol/wallet/actions/runs/14318379046/) (billed time 17m 53s)
Build once + container: [2m 54s](https://github.com/oasisprotocol/wallet/actions/runs/14319020422) (billed time 11m 36s)

![image](https://github.com/user-attachments/assets/af6be06d-d5e1-41bd-a0aa-8f4b617a0e3b)
.
![image](https://github.com/user-attachments/assets/b106ff12-1274-4c1e-98e4-628eea60116b)